Move integration tests to Integration Test module
