--- country-a-service/epps-service/src/test/java/dk/nsp/epps/service/IntegrationTests.java
@@ -1,76 +1,0 @@
-package dk.nsp.epps.service;
-
-import dk.dkma.medicinecard.xml_schema._2015._06._01.AuthorisedHealthcareProfessionalType;
-import dk.dkma.medicinecard.xml_schema._2015._06._01.DosageStructureForRequestType;
-import dk.dkma.medicinecard.xml_schema._2015._06._01.DosageStructuresForRequestType;
-import dk.dkma.medicinecard.xml_schema._2015._06._01.GetPrescriptionRequestType;
-import dk.dkma.medicinecard.xml_schema._2015._06._01.ModificatorPersonType;
-import dk.dkma.medicinecard.xml_schema._2015._06._01.ModificatorType;
-import dk.dkma.medicinecard.xml_schema._2015._06._01.OrganisationType;
-import dk.dkma.medicinecard.xml_schema._2015._06._01.PersonIdentifierType;
-import dk.dkma.medicinecard.xml_schema._2015._06._01.PredefinedOrganisationTypeType;
-import dk.dkma.medicinecard.xml_schema._2015._06._01.e2.CreateDrugMedicationRequestType;
-import dk.dkma.medicinecard.xml_schema._2015._06._01.e2.CreateDrugMedicationType;
-import dk.dkma.medicinecard.xml_schema._2015._06._01.e2.CreatePharmacyEffectuationOnPrescriptionType;
-import dk.dkma.medicinecard.xml_schema._2015._06._01.e2.CreatePharmacyEffectuationRequestType;
-import dk.dkma.medicinecard.xml_schema._2015._06._01.e2.CreatePrescriptionRequestType;
-import dk.dkma.medicinecard.xml_schema._2015._06._01.e2.GetMedicineCardRequestType;
-import dk.dkma.medicinecard.xml_schema._2015._06._01.e5.StartEffectuationRequestType;
-import dk.dkma.medicinecard.xml_schema._2015._06._01.e6.PrescriptionType;
-import dk.nsp.epps.client.CprClient;
-import dk.nsp.epps.client.FmkClient;
-import dk.nsp.epps.client.TestIdentities;
-import dk.nsp.epps.service.mapping.DispensationMapper;
-import dk.nsp.test.idp.EmployeeIdentities;
-import dk.nsp.epps.testing.shared.Cpr;
-import dk.nsp.epps.testing.shared.Fmk;
-import dk.nsp.test.idp.OrganizationIdentities;
-import dk.sdsd.dgws._2010._08.PredefinedRequestedRole;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Test;
-import org.w3c.dom.Document;
-
-import java.math.BigDecimal;
-import java.time.LocalDate;
-import java.time.ZonedDateTime;
-import java.util.Collection;
-
-/**
- * Integration tests for FMK and CPR.
- * To be run manually (for now) because they require external dependencies (including NSP STS).
- */
-@Disabled("Requires external dependencies")
-public class IntegrationTests {
-    private static final dk.dkma.medicinecard.xml_schema._2015._06._01.ObjectFactory medCardFac =
-        new dk.dkma.medicinecard.xml_schema._2015._06._01.ObjectFactory();
-
-
-    @Test
-    public void fmkGetPrescriptionAlternativeCaller() throws Exception {
-        var getPrescriptionRequest = GetPrescriptionRequestType.builder()
-            .withPersonIdentifier().withSource("CPR").withValue("1111111118").end()
-            .withIncludeOpenPrescriptions().end()
-            .build();
-
-        var prescriptions = Fmk.apiClient().getPrescription(getPrescriptionRequest, TestIdentities.apotekerJeppeMoeller);
-        Assertions.assertEquals("Helle", prescriptions.getPatient().getPerson().getName().getGivenName());
-    }
-
-
-    @Test
-    void cprGetPersonInformation() throws Exception {
-        var response = Cpr.apiClient().getPersonInformation("0611809735", OrganizationIdentities.sundhedsdatastyrelsen());
-        Assertions.assertEquals("Charles Test Babbage", response.getPersonInformationStructure()
-            .getRegularCPRPerson().getPersonNameForAddressingName());
-    }
-
-    @Test
-    void cprGetPersonInformationAlternativeCaller() throws Exception {
-        var response = Cpr.apiClient().getPersonInformation("0611809735", TestIdentities.apotekerJeppeMoeller);
-        Assertions.assertEquals("Charles Test Babbage", response.getPersonInformationStructure()
-            .getRegularCPRPerson().getPersonNameForAddressingName());
-    }
-
-
-}
